Refactor admin's features tests
